display error message in ErrorBoundary
