+++ client/src/_shared/index.js
@@ -1,3 +1,4 @@
 export { AuthLayout } from './AuthLayout'
 export { LoggedInLayout } from './LoggedInLayout'
 export { Header } from './Header'
+export { ErrorDisplay } from './ErrorDisplay'
